requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
cachetools==7.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
dj-database-url==3.1.0
Django==5.2.11
djangorestframework==3.16.1
gunicorn==25.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibm-cos-sdk==2.14.3
ibm-cos-sdk-core==2.14.3
ibm-cos-sdk-s3transfer==2.14.3
ibm_watsonx_ai==1.5.1
idna==3.11
jiter==0.13.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-core==1.2.11
langchain-deepseek==1.0.1
langchain-ibm==1.0.4
langchain-openai==1.1.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
lomond==0.3.3
numpy==2.4.2
openai==2.20.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tabulate==0.9.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
xxhash==3.6.0
zstandard==0.25.0
